split merge hander #1139
